Refactoring!
Brought it much closer to the sources as they are in T-Flows.
It works the same as the previous version, but all solvers
performs much faster, like two to three times faster.
That is very strange, maybe something is wrong with timing.

One solver, LDL^T, gives different residual.  In the old version
residual is of the order 10^-155, whereas here is 10^-16.  I did
a cereful check of all the sources and all seems fine.  I dunno.