IMPORTANT: RESOLVED A VERY SUBTLE BUG
A very unpleasant bug, and I suspect that something is wrong
with the Fortran compiler, I was using gnu.  To put it simply,
the initialization of local variables of the form:

logical :: transposed = .false.

was not working, if tranposed would be set later to the value
of optinal parameters passed.  Very strange behaviour.

On branch master

modified:   Solvers_Mod/Dense/Backward_Substitution.f90
modified:   Solvers_Mod/Dense/Forward_Substitution.f90